Moved 'functions' package to 'db/functions'
